modify benchmark module
