+++ app-benchmark/build.gradle.kts
@@ -4,7 +4,7 @@ plugins {
 }
 
 android {
-    namespace = "com.keelim.comssa.benchmark"
+    namespace = "com.keelim.benchmark"
     compileSdk = 33
 
     defaultConfig {
@@ -18,14 +18,32 @@ android {
         // This benchmark buildType is used for benchmarking, and should function like your
         // release build (for example, with minification on). It"s signed with a debug key
         // for easy local/CI testing.
-        create("app-comssa-benchmark") {
+        create("my-grade-benchmark") {
+            isDebuggable = true
+            signingConfig = getByName("debug").signingConfig
+            matchingFallbacks += listOf("release")
+        }
+        create("comssa-benchmark") {
+            isDebuggable = true
+            signingConfig = getByName("debug").signingConfig
+            matchingFallbacks += listOf("release")
+        }
+        create("nanda-benchmark") {
+            isDebuggable = true
+            signingConfig = getByName("debug").signingConfig
+            matchingFallbacks += listOf("release")
+        }
+        create("cnubus-benchmark") {
             isDebuggable = true
             signingConfig = getByName("debug").signingConfig
             matchingFallbacks += listOf("release")
         }
     }
 
+    targetProjectPath = ":app-my-grade"
     targetProjectPath = ":app-comssa"
+    targetProjectPath = ":app-nanda"
+    targetProjectPath = ":app-cnubus"
     experimentalProperties["android.experimental.self-instrumenting"] = true
 }
 
@@ -38,6 +56,9 @@ dependencies {
 
 androidComponents {
     beforeVariants(selector().all()) {
-        it.enabled = it.buildType == "app-comssa-benchmark"
+        it.enabled = (it.buildType == "my-grade-benchmark") ||
+            (it.buildType == "comssa-benchmark") ||
+            (it.buildType == "nanda-benchmark") ||
+            (it.buildType == "cnubus-benchmark")
     }
 }
