+++ app-benchmark/src/main/AndroidManifest.xml
@@ -2,6 +2,6 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android">
 
     <queries>
-        <package android:name="com.keelim.comssa" />
+        <package android:name="com.keelim.app.benchmark" />
     </queries>
 </manifest>
