--- app-cnubus-benchmark/src/main/AndroidManifest.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<manifest xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <queries>
-        <package android:name="com.keelim.cnubus" />
-    </queries>
-</manifest>
