--- app-cnubus-benchmark/src/main/java/com/keelim/benchmark/BaselineProfileGenerator.kt
@@ -1,25 +1,0 @@
-package com.keelim.benchmark
-
-import androidx.benchmark.macro.ExperimentalBaselineProfilesApi
-import androidx.benchmark.macro.junit4.BaselineProfileRule
-import androidx.test.ext.junit.runners.AndroidJUnit4
-import org.junit.Rule
-import org.junit.Test
-import org.junit.runner.RunWith
-
-@ExperimentalBaselineProfilesApi
-@RunWith(AndroidJUnit4::class)
-class BaselineProfileGenerator {
-    @get:Rule
-    val baselineProfileRule = BaselineProfileRule()
-
-    @Test
-    fun startup() =
-        baselineProfileRule.collectBaselineProfile(packageName = "com.keelim.cnubus") {
-            pressHome()
-            // This block defines the app's critical user journey. Here we are interested in
-            // optimizing for app startup. But you can also navigate and scroll
-            // through your most important UI.
-            startActivityAndWait()
-        }
-}
