+++ app-comssa/src/main/res/navigation/mobile_navigation.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/homeFragment"
-        android:name="com.keelim.comssa.ui.home.HomeFragment"
+        android:name="com.keelim.comssa.ui.screen.home.HomeFragment"
         android:label="fragment_home"
         tools:layout="@layout/fragment_home">
     </fragment>
