--- app-yr-benchmark/build.gradle.kts
@@ -1,43 +1,0 @@
-plugins {
-    id("com.android.test")
-    id("org.jetbrains.kotlin.android")
-}
-
-android {
-    namespace = "com.keelim.yr.benchmark"
-    compileSdk = 33
-
-    defaultConfig {
-        minSdk = 24
-        targetSdk = 33
-
-        testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
-    }
-
-    buildTypes {
-        // This benchmark buildType is used for benchmarking, and should function like your
-        // release build (for example, with minification on). It"s signed with a debug key
-        // for easy local/CI testing.
-        create("app-yr-benchmark") {
-            isDebuggable = true
-            signingConfig = getByName("debug").signingConfig
-            matchingFallbacks += listOf("release")
-        }
-    }
-
-    targetProjectPath = ":app-yr"
-    experimentalProperties["android.experimental.self-instrumenting"] = true
-}
-
-dependencies {
-    implementation("androidx.test.ext:junit:1.1.5")
-    implementation("androidx.test.espresso:espresso-core:3.5.1")
-    implementation("androidx.test.uiautomator:uiautomator:2.2.0")
-    implementation("androidx.benchmark:benchmark-macro-junit4:1.1.1")
-}
-
-androidComponents {
-    beforeVariants(selector().all()) {
-        it.enabled = it.buildType == "app-yr-benchmark"
-    }
-}
